frontend/app/src/features/auth/hooks/useCheckLoggedInUser.ts
import getCurrentUser from "features/auth/api/currentUserApi";
import useMessage from "hooks/useMessage";
import { useState } from "react";
import { useNavigate } from "react-router-dom";

const useCheckLoggedInUser = () => {
  const { showMessage } = useMessage();
  const [loading, setLoading] = useState(true);
  const navigate = useNavigate();
  const handleCheckLoggedInUser = async () => {
    try {
      const res = await getCurrentUser();
      if (!res.data.currentUser) {
        showMessage({ title: "ログインしてください。", status: "error" });
        navigate("/login");
      }
    } catch (e) {
      showMessage({ title: "エラーが発生しました。", status: "error" });
      navigate("/login");
    } finally {
      setLoading(false);
    }
  };
  return { handleCheckLoggedInUser, loading };
};

export default useCheckLoggedInUser;
